Fix excess key events on machines where long is 64bit
git-svn-id: svn://svn.xara.com/Trunk/XaraLX@1339 c9fdbc50-1c0f-0410-8134-ac6df33ab1bd

+++ wxOil/camelot.cpp
@@ -254,7 +254,7 @@ CCamApp::CCamApp()
 **********************************************************************************************/ 
 int /*TYPENOTE: Correct*/ CCamApp::FilterEvent( wxEvent& event )
 {
-	static INT32	lLastTimeStamp = 0;
+	static /*TYPENOTE: Correct*/ long	lLastTimeStamp = 0;
 
 	if (( event.GetEventType() == wxEVT_CREATE ) && (event.GetEventObject()->IsKindOf(CLASSINFO(wxTopLevelWindow))))
 	{
